Moved Entity class and repositories to posts package
