+++ posts/src/main/java/com/collreach/posts/model/bo/posts/Messages.java
@@ -1,4 +1,4 @@
-package com.collreach.posts.model.bo;
+package com.collreach.posts.model.bo.posts;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
