--- posts/src/main/java/com/collreach/posts/model/repositories/SeenRepository.java
@@ -1,8 +1,0 @@
-package com.collreach.posts.model.repositories;
-
-import com.collreach.posts.model.bo.Seen;
-import com.collreach.posts.model.bo.UserMessageKey;
-import org.springframework.data.repository.CrudRepository;
-
-public interface SeenRepository extends CrudRepository<Seen, UserMessageKey> {
-}
