+++ posts/src/main/java/com/collreach/posts/model/repositories/posts/MessagesRepository.java
@@ -1,6 +1,6 @@
-package com.collreach.posts.model.repositories;
+package com.collreach.posts.model.repositories.posts;
 
-import com.collreach.posts.model.bo.Messages;
+import com.collreach.posts.model.bo.posts.Messages;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MessagesRepository extends CrudRepository<Messages, Integer> {
