+++ posts/src/main/java/com/collreach/posts/model/repositories/posts/UsersRepository.java
@@ -1,6 +1,6 @@
-package com.collreach.posts.model.repositories;
+package com.collreach.posts.model.repositories.posts;
 
-import com.collreach.posts.model.bo.Users;
+import com.collreach.posts.model.bo.posts.Users;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UsersRepository extends CrudRepository<Users, Integer> {
